requirements.txt
asgiref==3.8.1
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
beautifulsoup4==4.12.3
certifi==2024.7.4
cffi==1.17.1
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.3.2
constantly==23.10.4
cryptography==43.0.1
daphne==4.1.2
distlib==0.3.8
dj-database-url==2.2.0
Django==3.2.1
django-cors-headers==4.3.1
django-environ==0.11.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
env==0.1.0
filelock==3.13.1
gunicorn==22.0.0
hyperlink==21.0.0
idna==3.7
incremental==24.7.2
msgpack==1.1.0
mysqlclient==2.2.5
packaging==24.1
pillow==10.4.0
platformdirs==4.2.0
psycopg2==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.8.0
pyOpenSSL==24.2.1
python-decouple==3.8
python-dotenv==1.0.1
pytz==2024.1
redis==5.0.8
requests==2.32.3
service-identity==24.1.0
setuptools==71.1.0
soupsieve==2.6
sqlparse==0.4.4
Twisted==24.7.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
virtualenv==20.25.1
whitenoise==6.7.0
zope.interface==7.0.3
